pyproject.toml
[project]
name = "llama-swap-setup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "pydantic>=2.11.7",
    "rich-click>=1.8.9",
    "structlog>=25.4.0",
    "ruamel.yaml>=0.18.6",
    "jinja2>=3.1.4",
    "pyyaml>=6.0.2",
    "pytest>=8.4.2",
    "requests>=2.32.5",
    "python-dotenv>=1.1.1",
    "tqdm>=4.66.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
